requirements.txt
streamlit==0.82.0
click==7.1.2
newspaper3k==0.2.8
lxml==4.6.3
nltk==3.6.2
pandas==1.1.3 
plotly==5.1.0
requests==2.25.1
beautifulsoup4==4.10.0
newsapi-python==0.2.6